Rename shared storage
