+++ upload-cli/src/index.ts
@@ -1,5 +1,5 @@
 import { ExifParserFactory } from 'ts-exif-parser';
-import { getFilesBy } from 'shared/src/storage';
+import { getFilesBy } from 'shared/storage';
 import { promises as fs } from 'fs';
 import { Storage } from '@google-cloud/storage';
 import dotenv from 'dotenv';
